-NF
-Created a main menu where you can click on the cards in front of you to access the feature listed on the card.
-Started laying groundwork for a functional deckbuilder array system that displays card text and establishes basic gameplay rules based on information provided by arrays of structs (containing a bool, a byte, and a text variable type).

Started things off by coding a working main menu. To get started on session hosting, I recommend you navigate to Blueprints->Cards+Widgets->Actor_HostGame_Child and work off of a "ActorOnClicked" event.

Trying to keep the file structure shallow for now to keep it from getting confusing. Naming conventions use the data type associated with the file (A struct will be S_NameHere, a game mode is GM_NameHere, a static mesh should be SM_NameHere—although I forgot to do that this time around). The function of the code in the blueprints should be reasonably documented. Let me know if any of it is confusing tomorrow.